Add files via upload
Among the many flaws and issues of constant debate in the United States today, are its healthcare system and stance on immigration. To examine the relationship between the two I used the 2017 American Community Survey data for Maryland, which contains a large range of social and demographic variables for persons and households in Maryland. I completed all analysis ad statistical testing through R.